% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnvi.R
\name{createEnvi}
\alias{createEnvi}
\title{Define and create a project environment}
\usage{
createEnvi(root_folder = tempdir(), folders = c("data", "data/tmp"),
  folder_names = NULL, path_prefix = "path_", global = FALSE,
  alternative_domain = NULL, alternative_root_folder = NULL)
}
\arguments{
\item{folders}{list of subfolders within the project root folder.}

\item{folder_names}{names of the variables that point to subfolders. If not
provided, the folder base paths of the folders is used.}

\item{path_prefix}{a prefix for the path variable names.}

\item{global}{export path strings as global variables.}

\item{alternative_domain}{alternative domain for which a different root path
than supplied by \code{root_folder} should be used.}

\item{alternative_root_folder}{alternative root folder for alternative domain.}

\item{root_folder_folder}{project root_folder folder directory.}
}
\value{
A list containing the project settings.
}
\description{
Defines folder structures and creates them if necessary, loads
libraries, and sets other project relevant parameters.
}
\examples{
\dontrun{
createEnvi(root_folder = tempdir(), folders = c("data/", "data/tmp/"))
}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
