% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createScript.R
\name{createScript}
\alias{createScript}
\title{Create files or scripts from templates}
\usage{
createScript(
  new_file = file.path(tempdir(), "tmp.R"),
  template = "script_function",
  notes = TRUE
)
}
\arguments{
\item{new_file}{name of the file to be created}

\item{template}{template to be used for the new file}

\item{notes}{logical: include notes from the template in the file}
}
\description{
Create files or scripts from brew templates supplied with the package.
}
\examples{
\dontrun{
createScript()
}

}
