% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnvi.R
\name{createEnvi}
\alias{createEnvi}
\title{Define and create a project environment}
\usage{
createEnvi(root_folder = tempdir(), folders = c("data", "data/tmp"),
  folder_names = NULL, path_prefix = "path_", global = FALSE,
  alternative_env_id = NULL, alternative_env_value = NULL,
  alternative_root_folder = NULL)
}
\arguments{
\item{root_folder}{project root folder directory.}

\item{folders}{list of subfolders within the project root folder.}

\item{folder_names}{names of the variables that point to subfolders. If not
provided, the folder base paths of the folders is used.}

\item{path_prefix}{a prefix for the path variable names.}

\item{global}{export path strings as global variables.}

\item{alternative_env_id}{alternative system environment attribute used to
check for setting an alternative \code{root_folder}.}

\item{alternative_env_value}{value of the attribute for which the alternative
\code{root_folder} should be set.}

\item{alternative_root_folder}{alternative root folder for alternative domain.}
}
\value{
A list containing the project settings.
}
\description{
Defines folder structures and creates them if necessary, loads
libraries, and sets other project relevant parameters.
}
\examples{
\dontrun{
createEnvi(root_folder = tempdir(), folders = c("data/", "data/tmp/"))
}
}
\author{
Christoph Reudenbach, Thomas Nauss
}
