% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpMa.R
\name{createScript}
\alias{createScript}
\title{Create files or scripts from templates}
\usage{
createScript(
  new_file = file.path(tempdir(), "tmp.R"),
  template = "script_function",
  notes = TRUE
)
}
\description{
Create files or scripts from brew templates supplied with the package.
}
\details{
None
}
\examples{
None
\dontrun{
createScript()
}

}
